requirements.txt
fastapi
uvicorn
tensorflow
numpy
pandas
keras
pillow
sqlalchemy
-e .